Major changes and organize into directories
Added 6 LEDs for backlit, a rotary encoder (EC11) and 0.91" OLED display. Changed the wiring from direct wiring to matrix wiring because I was running short on pins in the microcontroller, needed one more pin for the rotary encoder. Made a basic PCB in the PCB editor, need to arrange it later.